requirements.txt
flask>=2.0.0
pandas>=1.3.0
numpy
scikit-learn>=1.0.0
joblib>=1.0.0
requests>=2.26.0
python-whois>=0.7.3
dnspython>=2.1.0
tldextract
seaborn>=0.11.0
matplotlib>=3.4.0